requirements.txt
# Velvet Research - Full Dependencies
# Render Starter Plan

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
aiosqlite==0.19.0

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Email (Phase 2)
# sendgrid==6.11.0

# For report generation (imports from science_wizard_project)
anthropic>=0.18.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
